requirements.txt
python-dotenv
requests
pandas
openpyxl
pyinstaller